refactor: extract tab constants and types from tab-manager-helpers
Split tab-manager-helpers.js into three focused modules:
- tab-constants.js: all constants (DRAG_THRESHOLD, PANEL_MIN_WIDTH, etc.)
- tab-types.js: WorkspaceTab class
- tab-manager-helpers.js: pure helper functions + re-exports for backward compat

Updated imports in 17 consumer files to use the most specific module.

Closes #260
